requirements.txt
setuptools
Django
flake8
pep8
pyflakes
django-timezone-field
django-localflavor
dj-database-url
